+++ src/error.rs
@@ -1,0 +1,24 @@
+use crate::parser::*;
+
+use core::fmt;
+
+#[derive(Debug)]
+pub enum RError {
+    VariableNotFound(String),
+    ParseFailure(pest::error::Error<Rule>),
+}
+
+impl RError {
+    fn as_str(&self) -> String {
+        match self {
+            RError::VariableNotFound(v) => format!("Variable '{}' not found", v.as_str()),
+            RError::ParseFailure(e) => format!("Parse failed: {:?}", e),
+        }
+    }
+}
+
+impl fmt::Display for RError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
+}
