+++ src/lib.rs
@@ -1,0 +1,9 @@
+#[macro_use]
+extern crate pest_derive;
+
+pub mod ast;
+pub mod builtins;
+pub mod error;
+pub mod lang;
+pub mod parser;
+pub mod utils;
